+++ test/entry.js
@@ -27,3 +27,4 @@ window.zipLib = require('../src/lib/zip.js');
 window.quicknoteLib = require('../src/lib/quicknote.js');
 window.extcommandLib = require('../src/lib/extcommand.js');
 window.kanbanCardLib = require('../src/modals/kanbancard.js');
+window.vaultSyncRunLib = require('../src/lib/vaultsyncrun.js');